requirements.txt
# Fantasy Football Predictor - Production Requirements

# Core ML and Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
# Optional: for scripts/fetch_historical_nflverse.py (2023/2024 prior-year data)
# nfl_data_py>=0.3.0

# Additional ML
xgboost>=1.7.0
lightgbm>=4.0.0
optuna>=3.0.0

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
schedule>=1.2.0
